+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - new `--hide` option, that allows to filter out nodes from the profile based on passed regex. Resources of the filtered
   node are added to a parent node (function).
+- sierra gas profiling; in order to profile sierra gas please make sure to run snforge test with `--tracked-resource` flag set to "sierra-gas"
 
 ## [0.8.1] - 2025-02-25