+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - profiler now accounts for steps when pc is outside the function area
+- profiler now accounts for syscalls in nodes without `CairoExecutionInfo`
 
 ## [0.9.0] - 2025-05-20