+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2025-07-18
+
 ### Changed
 
 - addresses and selectors are now displayed in fixed hex format (66 chars) instead of decimal format